June/services/june-orchestrator/requirements.txt
# Existing dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
google-generativeai==0.7.2

# LiveKit SDK (NEW)
livekit==0.10.0
livekit-api==0.5.0

# Audio Processing (keep these)
numpy==1.24.3
soundfile==0.12.1